% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiAnovaOneW.h.R
\name{aiAnovaOneW}
\alias{aiAnovaOneW}
\title{One-Way ANOVA + Report}
\usage{
aiAnovaOneW(
  data,
  deps,
  group,
  fishers = TRUE,
  welchs = FALSE,
  kruskal = FALSE,
  phMethod = "tukey",
  desc = TRUE,
  norm = TRUE,
  eqv = TRUE,
  useLLM = FALSE,
  backend = "auto",
  interp = TRUE,
  polish = FALSE,
  model = "qwen3.5:4b",
  endpoint = ""
)
}
\arguments{
\item{data}{.}

\item{deps}{.}

\item{group}{.}

\item{fishers}{.}

\item{welchs}{.}

\item{kruskal}{.}

\item{phMethod}{.}

\item{desc}{.}

\item{norm}{.}

\item{eqv}{.}

\item{useLLM}{.}

\item{backend}{.}

\item{interp}{.}

\item{polish}{.}

\item{model}{.}

\item{endpoint}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$anova} \tab \tab \tab \tab \tab a table \cr
  \code{results$kw} \tab \tab \tab \tab \tab a table \cr
  \code{results$desc} \tab \tab \tab \tab \tab a table \cr
  \code{results$norm} \tab \tab \tab \tab \tab a table \cr
  \code{results$eqv} \tab \tab \tab \tab \tab a table \cr
  \code{results$postHoc} \tab \tab \tab \tab \tab a html \cr
  \code{results$report} \tab \tab \tab \tab \tab a html \cr
  \code{results$warnings} \tab \tab \tab \tab \tab a html \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$anova$asDF}

\code{as.data.frame(results$anova)}
}
\description{
One-way ANOVA (Fisher / Welch) and Kruskal-Wallis with post hoc tests and
an APA-style report paragraph.
}
