% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{summarize_results}
\alias{summarize_results}
\title{Summarise any jmv results object}
\usage{
summarize_results(res)
}
\arguments{
\item{res}{A jmv results object.}
}
\description{
Summarise any jmv results object
}
